Matlab/Simulink Model of a Power Electronic Converter for Hydrogen Production via Water Electrolysis
Context: A successful climate protection requires a shift away from fossil energy to renewable energy. Hydrogen already plays a role in today's fossil energy system as a secondary energy carrier for refinery processes and the chemical industry. In principle, hydrogen is technically conceivable as a final energy carrier in all areas of application, such as transportation, industry and buildings. However, hydrogen must be low-carbon from the outset and ultimately be environmentally friendly produced by electrolysis with renewable energy.

Objective: In this project, a Matlab/Simulink model of a power supply for a 1 𝑀𝑊 water electrolyzer is to be developed. The connection to the medium voltage grid is made via an MV transformer. An active, three-phase converter is used to generate an adjustable DC voltage. A buck converter with an inter- leaved topology is used for DC current regulation. The characteristic of the electrolyzer is described by its polarization curve.

Procedure: Develop a simulation model in Simulink using the “Simscape” li- brary to model all electrical components. In particular, the library “Special- ized Power Systems” is to be used. Created the model step by step from the grid connection via transformer, active rectifier, buck converter and electrolyser model. Modeled the load behavior by means of an equivalent voltage source. Parameterized all components, the associated values are saved in a parameter file (m-file).
Used the following specification:
Medium-voltage grid 6 𝐾𝑉
Rectifier input filter inductance 𝐿𝑖𝑛 = 1 𝑚𝐻
Switching frequency of the IGBT rectifier 𝑓𝑆,𝑟𝑒𝑐 = 4 𝑘𝐻𝑧. DC link capacitance 𝐶 = 2400 𝜇𝐹
DC link voltage 𝑉𝐷𝐶 = 800 𝑉
Switching frequency of the buck converter 𝑓𝑆,𝑏𝑢𝑐𝑘 = 2 𝑘𝐻𝑧

Determined the inductance 𝐿𝐷𝐶 of one buck converter stage so that the electrolyzer DC current 𝐼𝐷𝐶 has a ripple of ∆𝑖𝐷𝐶 = 200 𝐴 at the nominal operating point (1 𝑀𝑊). The IGBT rectifier (3-phase converter) is to be controlled so that a specified reactive power 𝑄𝑔𝑟𝑖𝑑,𝑟𝑒𝑓 = 0 is drawn from the grid and that the output voltage is regulated to a setpoint value 𝑉𝐷𝐶,𝑟𝑒𝑓 = 360 𝑉. For this purpose, the control method from lecture “Power Control of Renewable Energy Systems” in chapter 8.3 is to be used, which is applied for the power control of a grid-side converter in a wind turbine. A suitable method from the literature should be applied to control the buck converter, e.g. Hartmut Hinz; Noass Kunstbergs “Control Methods for a Bi-directional DC-DC Converter in Fuel Cell Hybrid Vehicle Applications,” 2021 IEEE 4th International Conference on Power and Energy Applications (ICPEA).